+++ src/CustomRateFilter.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+class GJGameLevel;
+
+namespace rating_filters {
+
+bool isRatedLevel(GJGameLevel* level);
+
+} // namespace rating_filters